fix: casting ricker source data type to numpy.float32

+++ geod67lib/wavefield/wavelet.py
@@ -4,4 +4,4 @@ def ricker(fpeak, dt, nt):
     t = np.arange(nt) * dt
     t0 = 1/fpeak
     t_ = t - t0
-    return (1 - 2 * np.pi**2 * fpeak**2 * t_**2) * np.exp(-np.pi**2 * fpeak**2 * t_**2)
+    return np.float32((1 - 2 * np.pi**2 * fpeak**2 * t_**2) * np.exp(-np.pi**2 * fpeak**2 * t_**2))
